refactor: 重构代码，删除无用代码，修改变量名、注释

+++ main/app/service/business.go
@@ -163,8 +163,8 @@ func (s *businessSrv) Printer(ctx context.Context, printerReq req.BusinessDataPr
 			PeakHourList: func() []business_data_resp.PeakHour {
 				peakHours, err := repository.NewSaleOrderPeakTimeRepo(ctx.GetDB()).GetMaxRecord(
 					storeSetting.TimeZone,
-					uint(printerParam.QueryStartTime),
-					uint(printerParam.QueryEndTime),
+					printerParam.QueryStartTime,
+					printerParam.QueryEndTime,
 					ctx.GetStaffUuid(),
 				)
 				if err != nil {
@@ -291,8 +291,8 @@ func (s *businessSrv) CountBusiness(ctx context.Context, req req.BusinessDataCou
 	// 销售数据
 	saleData := s.statisticsSrv.CountSale(ctx, CountReq{
 		TimeType:       req.TimeType,
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 		CategoryType:   req.CategoryType,
 		DutyNo:         req.DutyNo,
 	})
@@ -309,14 +309,14 @@ func (s *businessSrv) CountBusiness(ctx context.Context, req req.BusinessDataCou
 	// 会员数量
 	memberNum := s.statisticsSrv.CountMemberNum(ctx, CountReq{
 		TimeType:       req.TimeType,
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 	})
 	// 未结订单
 	unpaidOrderData := s.statisticsSrv.CountUnpaidOrder(ctx, CountReq{
 		TimeType:       req.TimeType,
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 		CategoryType:   req.CategoryType,
 		DutyNo:         req.DutyNo,
 	})
@@ -375,8 +375,8 @@ func (s *businessSrv) CountBusiness(ctx context.Context, req req.BusinessDataCou
 		MemberData: func() business_data_resp.MemberData {
 			memberData := s.statisticsSrv.CountMember(ctx, CountReq{
 				TimeType:       req.TimeType,
-				QueryStartTime: int64(req.QueryStartTime),
-				QueryEndTime:   int64(req.QueryEndTime),
+				QueryStartTime: req.QueryStartTime,
+				QueryEndTime:   req.QueryEndTime,
 				CategoryType:   req.CategoryType,
 			})
 			return business_data_resp.MemberData{
@@ -410,8 +410,8 @@ func (s *businessSrv) CountBusiness(ctx context.Context, req req.BusinessDataCou
 		PercentageList: func() []business_data_resp.Percentage {
 			taxData := s.statisticsSrv.CountTax(ctx, CountReq{
 				TimeType:       req.TimeType,
-				QueryStartTime: int64(req.QueryStartTime),
-				QueryEndTime:   int64(req.QueryEndTime),
+				QueryStartTime: req.QueryStartTime,
+				QueryEndTime:   req.QueryEndTime,
 				CategoryType:   req.CategoryType,
 			})
 			list := make([]business_data_resp.Percentage, 0, len(taxData))
@@ -485,8 +485,8 @@ func (s *businessSrv) CountProduct(ctx context.Context, req req.BusinessDataCoun
 		Products: func() []business_data_resp.Product {
 			productList := s.statisticsSrv.CountProduct(ctx, CountReq{
 				TimeType:       req.TimeType,
-				QueryStartTime: int64(req.QueryStartTime),
-				QueryEndTime:   int64(req.QueryEndTime),
+				QueryStartTime: req.QueryStartTime,
+				QueryEndTime:   req.QueryEndTime,
 				CategoryType:   req.CategoryType,
 				DutyNo:         req.DutyNo,
 			})
@@ -511,8 +511,8 @@ func (s *businessSrv) CountProduct(ctx context.Context, req req.BusinessDataCoun
 func (s *businessSrv) CountArea(ctx context.Context, req req.BusinessDataCountReq) (*business_data_resp.BusinessDataArea, error) {
 	areaData := s.statisticsSrv.CountArea(ctx, CountReq{
 		TimeType:       req.TimeType,
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 		CategoryType:   req.CategoryType,
 		DutyNo:         req.DutyNo,
 	})
@@ -538,8 +538,8 @@ func (s *businessSrv) RankProduct(ctx context.Context, req req.BusinessDataRankP
 		Ranks: func() []business_data_resp.ProductRank {
 			productRankList := s.statisticsSrv.RankProduct(ctx, CountReq{
 				RankType:       req.RankType,
-				QueryStartTime: int64(req.QueryStartTime),
-				QueryEndTime:   int64(req.QueryEndTime),
+				QueryStartTime: req.QueryStartTime,
+				QueryEndTime:   req.QueryEndTime,
 			})
 			list := make([]business_data_resp.ProductRank, 0, len(productRankList))
 			for _, productRank := range productRankList {
@@ -598,8 +598,8 @@ func (s *businessSrv) CountProductSales(ctx context.Context, req req.BusinessDat
 // Count7Days 统计7天
 func (s *businessSrv) Count7Days(ctx context.Context, req req.BusinessDataCountReq) (*business_data_resp.BusinessDataCount7Days, error) {
 	sevenDaysData := s.statisticsSrv.Count7Days(ctx, CountReq{
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 		Timezone:       ctx.GetCompany().CompanySetting.Timezone,
 	})
 
@@ -780,8 +780,8 @@ func (s *businessSrv) CountHome(ctx context.Context, req req.BusinessDataCountRe
 	// 销售数据
 	saleData := s.statisticsSrv.CountSale(ctx, CountReq{
 		TimeType:       req.TimeType,
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 		CategoryType:   req.CategoryType,
 		DutyNo:         req.DutyNo,
 	})
@@ -789,8 +789,8 @@ func (s *businessSrv) CountHome(ctx context.Context, req req.BusinessDataCountRe
 	// 会员数量
 	memberNum := s.statisticsSrv.CountMemberNum(ctx, CountReq{
 		TimeType:       req.TimeType,
-		QueryStartTime: int64(req.QueryStartTime),
-		QueryEndTime:   int64(req.QueryEndTime),
+		QueryStartTime: req.QueryStartTime,
+		QueryEndTime:   req.QueryEndTime,
 	})
 
 	// 首页数据
@@ -803,8 +803,8 @@ func (s *businessSrv) CountHome(ctx context.Context, req req.BusinessDataCountRe
 		MemberData: func() business_data_resp.MemberData {
 			memberData := s.statisticsSrv.CountMember(ctx, CountReq{
 				TimeType:       req.TimeType,
-				QueryStartTime: int64(req.QueryStartTime),
-				QueryEndTime:   int64(req.QueryEndTime),
+				QueryStartTime: req.QueryStartTime,
+				QueryEndTime:   req.QueryEndTime,
 				CategoryType:   req.CategoryType,
 			})
 			return business_data_resp.MemberData{
